refactor: 프로젝트 notice 컴포넌트들 components>project>notice 이하로 경로 이동
