+++ src/app/project/notice/layout.tsx
@@ -1,5 +1,5 @@
 import React, {ReactNode} from 'react';
-import NoticeNavTab from "@/components/notice/NoticeNavTab";
+import NoticeNavTab from "@/components/project/notice/NoticeNavTab";
 
 function NoticeLayout({children}:{children:ReactNode}) {
     return (
